chore: Rename subgraphUtils to subgraph-utils file
